Add a utility class to hide elements until they are upgraded
An emerging pattern is to have elements which should be hidden until the
containing element is either upgraded or JS load times out or fails.

Adding a `u-js-hide-while-loading` class to any element will result in
it being hidden on page load in JS-capable user agents until the
containing element is successfully upgraded, at which point the class is
removed.

Doing this with a class on the element enables this mechanism to work on
a component rather than whole-document level.